Add rubocop

+++ generate.rb
@@ -1,9 +1,10 @@
 #! /usr/bin/env ruby
+# frozen_string_literal: true
 
 require 'pathname'
 require 'active_support/core_ext/string'
 
-dir = File.expand_path(File.dirname(__FILE__))
+dir = __dir__
 Dir.chdir dir
 
 args = ARGV
@@ -23,53 +24,57 @@ year = args[0]
 day = args[1].to_s.rjust(2, '0')
 
 filename = args[2] || begin
-  page = `curl -sS --cookie "session=#{$session}" -XGET https://adventofcode.com/#{year}/day/#{day.to_i}`
-  filename = page[/<h2[^>]*>.*Day\s*\d+:(.*)\s-*<\/h2>/, 1]
+  page = `curl -sS --cookie "session=#{@session}" -XGET https://adventofcode.com/#{year}/day/#{day.to_i}`
+  filename = page[%r{<h2[^>]*>.*Day\s*\d+:(.*)\s-*</h2>}, 1]
 end
 
-fail 'Cannot find problem name' if filename.nil?
+raise 'Cannot find problem name' if filename.nil?
 
 filename = filename.strip.gsub(/[^A-Za-z]/, '').underscore
 
 lib_file = "./#{year}/lib/#{day}_#{filename}.rb"
-File.open(lib_file, 'w') do |f|
-  f.write <<~LIB
-    class #{filename.camelcase}
-      def initialize(input)
-        @input = input
-      end
+unless File.exist?(lib_file)
+  File.open(lib_file, 'w') do |f|
+    f.write <<~LIB
+      class #{filename.camelcase}
+        def initialize(input)
+          @input = input
+        end
 
-      def part1
-        true
-      end
+        def part1
+          true
+        end
 
-      def part2
-        true
+        def part2
+          true
+        end
       end
-    end
-  LIB
-end unless File.exist?(lib_file)
+    LIB
+  end
+end
 
 test_file = "./#{year}/test/#{day}.rb"
-File.open(test_file, 'w') do |f|
-  f.write <<~TEST
-    require 'minitest/autorun'
-    require_relative('../lib/#{day}_#{filename}.rb')
+unless File.exist?(test_file)
+  File.open(test_file, 'w') do |f|
+    f.write <<~TEST
+      require 'minitest/autorun'
+      require_relative('../lib/#{day}_#{filename}.rb')
 
-    describe #{filename.camelcase} do
-      before { @k = #{filename.camelcase} }
+      describe #{filename.camelcase} do
+        before { @k = #{filename.camelcase} }
 
-      def test_part1
-        assert_equal true, @k.new(<<~INPUT).part1
-        INPUT
+        def test_part1
+          assert_equal true, @k.new(<<~INPUT).part1
+          INPUT
+        end
+
+        def test_part2
+          assert_equal true, @k.new(<<~INPUT).part2
+          INPUT
+        end
       end
-
-      def test_part2
-        assert_equal true, @k.new(<<~INPUT).part2
-        INPUT
-      end
-    end
-  TEST
-end unless File.exist?(test_file)
+    TEST
+  end
+end
 
 `$EDITOR #{lib_file} #{test_file}`
